pusg

+++ routes/cashier_route.py
@@ -177,25 +177,28 @@ async def expance(current_user = current_user_dep, database = database_dep):
 
 @app.post("/delay/check")
 async def delay_check(check_id:int, db = database_dep):
-    obj = db.query(models.Sale).filter(models.Sale.id == check_id).first()
-    for i in obj.items:
-        product = db.query(models.Product).filter(models.Product.id == i.product_id).first()
-        box = 0
-        package = 0
-        from_package = 0
-        if i.amount_of_box:
-            box = product.amount_in_box *  product.amount_in_package * i.amount_of_box 
-        if i.amount_of_package:
-            package = product.amount_in_package * i.amount_of_package
-        if i.amount_from_package:
-            from_package = i.amount_from_package
-        product.overall_amount += sum([box, package,from_package])
-        print([box, package,from_package])
-        if box:
-            product.box += i.amount_of_box
-        db.delete(i)
+    if check_id:
+        obj = db.query(models.Sale).filter(models.Sale.id == check_id).first()
+        for i in obj.items:
+            product = db.query(models.Product).filter(models.Product.id == i.product_id).first()
+            box = 0
+            package = 0
+            from_package = 0
+            if i.amount_of_box:
+                box = product.amount_in_box *  product.amount_in_package * i.amount_of_box 
+            if i.amount_of_package:
+                package = product.amount_in_package * i.amount_of_package
+            if i.amount_from_package:
+                from_package = i.amount_from_package
+            product.overall_amount += sum([box, package,from_package])
+            print([box, package,from_package])
+            if box:
+                product.box += i.amount_of_box
+            db.delete(i)
+            db.commit()
+
+        db.delete(obj)
         db.commit()
-    
-    db.delete(obj)
-    db.commit()
-    return {"message":"success"}
+        return {"message":"success"}
+    else:
+        return {"message":"Queriyda check id yo'q"}
